+++ src/main/java/persistence/member/MemberOutflowMapper.java
@@ -2,13 +2,12 @@ package persistence.member;
 
 import domain.member.MemberOutflow;
 import domain.member.MemberOutflowExample;
+import java.util.List;
 import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.session.RowBounds;
 
-import java.util.List;
-
 public interface MemberOutflowMapper {
-    int countByExample(MemberOutflowExample example);
+    long countByExample(MemberOutflowExample example);
 
     int deleteByExample(MemberOutflowExample example);
