+++ src/main/java/service/member/MemberInflowOutService.java
@@ -66,7 +66,7 @@ public class MemberInflowOutService extends MemberBaseMapper {
         if(partyId!=null) criteria.andPartyIdEqualTo(partyId);
         if(branchId!=null) criteria.andBranchIdEqualTo(branchId);
 
-        return memberInflowMapper.countByExample(example);
+        return (int) memberInflowMapper.countByExample(example);
     }
 
     // 上一个 （查找比当前记录的“创建时间”  小  的记录中的  最大  的“创建时间”的记录）
